refactor(root): delegate query and diagnostics to owners
Move admitted graph-query adapters, branch diagnostics, store-meta
counters, and fact-owner identity into their owner crates so TraceDecay
keeps handles and typed results instead of façade logic.
