+++ crates/tracedecay/src/mcp/tools/handlers/dispatch_test_support.rs
@@ -228,7 +228,7 @@ fn verified_graph_options_with_freshness<'a>(
     options.admitted_project_scope = Some(scope.clone());
     options.code_graph_read_admission_port = Some(Arc::new(FixtureCodeGraphAdmission { scope }));
     options.verified_graph_query_port = Some(
-        crate::tracedecay::queries::graph::admitted_verified_graph_query_port_with_source(
+        tracedecay_graph_query::admitted_verified_graph_query_port_with_source(
             options
                 .code_graph_read_admission_port
                 .clone()
@@ -256,8 +256,8 @@ pub(super) fn verified_graph_error_options<'a>(
     let mut options = verified_graph_options(cg, options);
     options.code_graph_projection_read_port =
         Some(Arc::new(FailingFixtureCodeGraphProjection { error }));
-    options.verified_graph_query_port = Some(
-        crate::tracedecay::queries::graph::admitted_verified_graph_query_port(
+    options.verified_graph_query_port =
+        Some(tracedecay_graph_query::admitted_verified_graph_query_port(
             options
                 .code_graph_read_admission_port
                 .clone()
@@ -266,8 +266,7 @@ pub(super) fn verified_graph_error_options<'a>(
                 .code_graph_projection_read_port
                 .clone()
                 .expect("graph fixture projection"),
-        ),
-    );
+        ));
     options
 }
