+++ crates/tracedecay/src/tracedecay.rs
@@ -11,6 +11,7 @@ use std::sync::{Arc, OnceLock};
 use crate::config::TraceDecayConfig;
 use tracedecay_contracts::context_scout::ContextScoutAddressV1;
 use tracedecay_domain::errors::Result;
+use tracedecay_graph_query::SourceReadContext;
 use tracedecay_runtime_core::db::{Database, DatabaseStorageTelemetryHandle};
 use tracedecay_runtime_core::storage::{self, StoreLayout};
 use tracedecay_store_runtime::DaemonSessionRuntimeRegistryV1;
@@ -57,7 +58,7 @@ pub struct TraceDecay {
     context_scout_owner: Option<
         Arc<tracedecay_agent_hosts::agents::context_scout_owner::ProjectContextScoutOwnerV1>,
     >,
-    context_scout_claim_authorities: tokio::sync::RwLock<Vec<MountedContextScoutClaimAuthorityV1>>,
+    context_scout_claim_authorities: tokio::sync::RwLock<Vec<MountedContextScoutClaim>>,
     #[cfg(any(test, feature = "test-transport"))]
     test_runtime_guard: Option<Arc<crate::host_admission::HostAdmissionTestRuntimeV1>>,
     _standalone_maintenance_scope:
@@ -66,17 +67,14 @@ pub struct TraceDecay {
 
 const MAX_MOUNTED_CONTEXT_SCOUT_CLAIM_AUTHORITIES: usize = 256;
 
-#[derive(Clone)]
-struct MountedContextScoutClaimAuthorityV1 {
-    registry: Arc<
-        tracedecay_agent_hosts::agents::context_scout_ports::ProjectContextScoutAddressRegistryV1,
-    >,
-    pin: tracedecay_agent_hosts::agents::context_scout_ports::ContextScoutAuthorityPinV1,
-    context: tracedecay_contracts::RequestContext,
-    lifecycle: tracedecay_agent_hosts::agents::context_scout_ports::ContextScoutLifecycleAddressV1,
-    address: ContextScoutAddressV1,
-    input_watermark: [u8; 32],
-}
+type MountedContextScoutClaim = (
+    Arc<tracedecay_agent_hosts::agents::context_scout_ports::ProjectContextScoutAddressRegistryV1>,
+    tracedecay_agent_hosts::agents::context_scout_ports::ContextScoutAuthorityPinV1,
+    tracedecay_contracts::RequestContext,
+    tracedecay_agent_hosts::agents::context_scout_ports::ContextScoutLifecycleAddressV1,
+    ContextScoutAddressV1,
+    [u8; 32],
+);
 
 impl TraceDecay {
     pub(crate) fn storage_telemetry_handle(&self) -> Result<DatabaseStorageTelemetryHandle> {
@@ -112,6 +110,15 @@ impl TraceDecay {
 
     pub(crate) fn hook_store_layout(&self) -> &StoreLayout {
         &self.store_layout
+    }
+
+    pub(crate) fn source_read_context(&self) -> Option<SourceReadContext> {
+        Some(SourceReadContext::new(
+            self.project_root.clone(),
+            self.db.clone(),
+            self.read_only,
+            self.store_layout.identity.project_id.clone()?,
+        ))
     }
 
     pub(crate) fn context_scout_owner(
@@ -150,18 +157,11 @@ impl TraceDecay {
         {
             return false;
         }
-        let mounted = MountedContextScoutClaimAuthorityV1 {
-            registry,
-            pin,
-            context,
-            lifecycle,
-            address,
-            input_watermark,
-        };
+        let mounted = (registry, pin, context, lifecycle, address, input_watermark);
         let mut authorities = self.context_scout_claim_authorities.write().await;
         if let Some(existing) = authorities
             .iter_mut()
-            .find(|existing| existing.lifecycle == mounted.lifecycle)
+            .find(|existing| existing.3 == mounted.3)
         {
             *existing = mounted;
             return true;
@@ -184,35 +184,27 @@ impl TraceDecay {
         lifecycle: &tracedecay_agent_hosts::agents::context_scout_ports::ContextScoutLifecycleAddressV1,
         observed_at: tracedecay_domain::UtcMicros,
     ) -> Option<(ContextScoutAddressV1, [u8; 32])> {
-        let mounted = self
+        let (registry, pin, context, _, address, input_watermark) = self
             .context_scout_claim_authorities
             .read()
             .await
             .iter()
-            .find(|mounted| mounted.lifecycle == *lifecycle)
+            .find(|mounted| mounted.3 == *lifecycle)
             .cloned()?;
-        if !self
-            .context_scout_configuration_is_current(&mounted.pin)
-            .await
-        {
+        if !self.context_scout_configuration_is_current(&pin).await {
             return None;
         }
-        let resolved = mounted
-            .registry
-            .resolve_current_exact(hook, &mounted.pin, lifecycle, &mounted.context, observed_at)
+        let resolved = registry
+            .resolve_current_exact(hook, &pin, lifecycle, &context, observed_at)
             .await;
         let resolved = (resolved
             == tracedecay_agent_hosts::agents::context_scout_ports::ContextScoutAddressResolveOutcomeV1::Resolved(
-                mounted.address,
+                address,
             ))
-        .then_some((mounted.address, mounted.input_watermark));
+        .then_some((address, input_watermark));
         // Re-check currentness after the registry read: a configuration
         // revision that lands mid-resolve must not hand out a stale claim.
-        if resolved.is_some()
-            && self
-                .context_scout_configuration_is_current(&mounted.pin)
-                .await
-        {
+        if resolved.is_some() && self.context_scout_configuration_is_current(&pin).await {
             resolved
         } else {
             None
