+++ crates/tracedecay/src/tracedecay/diagnostics.rs
@@ -1,11 +1,10 @@
 //! Branch-state accessors and branch-tracking diagnostics for the open
 //! store.
 
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
 use tracedecay_domain::errors::{Result, TraceDecayError};
 use tracedecay_runtime_core::branch;
-use tracedecay_runtime_core::branch_meta;
 use tracedecay_runtime_core::db::Database;
 use tracedecay_runtime_core::storage::StoreLayout;
 
@@ -152,196 +151,8 @@ impl TraceDecay {
         Ok(database)
     }
 
-    fn build_branch_diagnostics(
-        project_root: &Path,
-        data_root: &Path,
-        open_active_branch: Option<String>,
-        serving_branch: Option<String>,
-        fallback_warning: Option<String>,
-        serving_db_path: PathBuf,
-    ) -> BranchDiagnostics {
-        let meta = branch_meta::load_branch_meta(data_root);
-        let current_branch = branch::current_branch(project_root);
-        let tracking_enabled = meta.as_ref().is_some_and(|m| !m.branches.is_empty());
-        let branch_drifted =
-            tracking_enabled && current_branch.as_deref() != open_active_branch.as_deref();
-        let is_fallback = fallback_warning.is_some();
-        let fallback_target = if is_fallback {
-            serving_branch.clone()
-        } else {
-            None
-        };
-        let serving_db_exists = serving_db_path.exists();
-
-        let (
-            live_branch_tracked,
-            live_branch_db_path,
-            live_branch_db_exists,
-            nearest_tracked_ancestor,
-            nearest_tracked_ancestor_db_path,
-            nearest_tracked_ancestor_db_exists,
-        ) = if let (Some(meta), Some(current)) = (meta.as_ref(), current_branch.as_deref()) {
-            let live_branch_tracked = meta.is_tracked(current);
-            let live_branch_db_path = if live_branch_tracked {
-                branch::resolve_branch_db_path(data_root, current, meta)
-            } else {
-                None
-            };
-            let live_branch_db_exists = live_branch_db_path.as_ref().map(|path| path.exists());
-            let nearest_tracked_ancestor = if live_branch_tracked {
-                None
-            } else {
-                branch::find_nearest_tracked_ancestor(project_root, current, meta)
-            };
-            let nearest_tracked_ancestor_db_path = nearest_tracked_ancestor
-                .as_deref()
-                .and_then(|ancestor| branch::resolve_branch_db_path(data_root, ancestor, meta));
-            let nearest_tracked_ancestor_db_exists = nearest_tracked_ancestor_db_path
-                .as_ref()
-                .map(|path| path.exists());
-            (
-                live_branch_tracked,
-                live_branch_db_path,
-                live_branch_db_exists,
-                nearest_tracked_ancestor,
-                nearest_tracked_ancestor_db_path,
-                nearest_tracked_ancestor_db_exists,
-            )
-        } else {
-            (false, None, None, None, None, None)
-        };
-
-        let mut warnings = Vec::new();
-        if branch_drifted {
-            warnings.push(format!(
-                "branch drift detected: working tree is on '{}' but this instance opened on '{}' and is still serving '{}'. Reopen the index so reads and writes target the live branch.",
-                current_branch.as_deref().unwrap_or("detached HEAD"),
-                open_active_branch.as_deref().unwrap_or("detached HEAD"),
-                serving_branch.as_deref().unwrap_or("default branch"),
-            ));
-        }
-        if !serving_db_exists {
-            warnings.push(format!(
-                "serving branch '{}' points at a missing DB: {}",
-                serving_branch.as_deref().unwrap_or("default branch"),
-                serving_db_path.display(),
-            ));
-        }
-        if let (Some(current), Some(false), Some(path)) = (
-            current_branch.as_deref(),
-            live_branch_db_exists,
-            live_branch_db_path.as_ref(),
-        ) {
-            warnings.push(format!(
-                "tracked branch '{}' is listed in branch metadata but its DB is missing at '{}'; serving '{}' instead.",
-                current,
-                path.display(),
-                serving_branch.as_deref().unwrap_or("default branch"),
-            ));
-        } else if is_fallback {
-            match (
-                current_branch.as_deref(),
-                nearest_tracked_ancestor.as_deref(),
-                fallback_target.as_deref(),
-            ) {
-                (Some(current), Some(ancestor), Some(target)) => warnings.push(format!(
-                    "branch '{current}' is not tracked; nearest indexed ancestor is '{ancestor}' and tracedecay is serving '{target}' instead."
-                )),
-                (Some(current), None, Some(target)) => warnings.push(format!(
-                    "branch '{current}' is not tracked and no indexed ancestor DB was available; tracedecay is serving '{target}' instead."
-                )),
-                _ => {}
-            }
-        }
-
-        let branch_resolution = if !tracking_enabled {
-            "single_db".to_string()
-        } else if branch_drifted {
-            "stale_serving_branch".to_string()
-        } else if current_branch.is_none() {
-            "detached_default".to_string()
-        } else if is_fallback {
-            match (
-                nearest_tracked_ancestor.as_deref(),
-                fallback_target.as_deref(),
-            ) {
-                (Some(ancestor), Some(target)) if ancestor == target => {
-                    "fallback_ancestor".to_string()
-                }
-                _ => "fallback_default".to_string(),
-            }
-        } else {
-            "exact".to_string()
-        };
-
-        let mut branches = Vec::new();
-        if let Some(meta) = meta.as_ref() {
-            let mut names: Vec<_> = meta.branches.keys().cloned().collect();
-            names.sort();
-            for name in names {
-                let entry = &meta.branches[&name];
-                let db_path = data_root.join(&entry.db_file);
-                let db_exists = db_path.exists();
-                let size_bytes = db_path.metadata().map_or(0, |metadata| metadata.len());
-                let parent_db_path = entry
-                    .parent
-                    .as_deref()
-                    .and_then(|parent| branch::resolve_branch_db_path(data_root, parent, meta));
-                let parent_db_exists = parent_db_path.as_ref().map(|path| path.exists());
-                let mut branch_warnings = Vec::new();
-                if !db_exists {
-                    branch_warnings.push(format!("missing DB at '{}'", db_path.display()));
-                }
-                if entry.parent.is_some() && parent_db_exists == Some(false) {
-                    branch_warnings.push("parent DB is missing".to_string());
-                }
-                branches.push(TrackedBranchDiagnostic {
-                    name: name.clone(),
-                    db_file: entry.db_file.clone(),
-                    db_path,
-                    db_exists,
-                    size_bytes,
-                    parent: entry.parent.clone(),
-                    parent_db_path,
-                    parent_db_exists,
-                    created_at: entry.created_at.clone(),
-                    last_synced_at: entry.last_synced_at.clone(),
-                    is_default: name == meta.default_branch,
-                    is_current: current_branch.as_deref() == Some(name.as_str()),
-                    is_open_active: open_active_branch.as_deref() == Some(name.as_str()),
-                    is_serving: serving_branch.as_deref() == Some(name.as_str()),
-                    warnings: branch_warnings,
-                });
-            }
-        }
-
-        BranchDiagnostics {
-            tracking_enabled,
-            default_branch: meta.as_ref().map(|m| m.default_branch.clone()),
-            current_branch,
-            open_active_branch,
-            serving_branch,
-            serving_db_path,
-            serving_db_exists,
-            branch_drifted,
-            branch_resolution,
-            is_fallback,
-            fallback_target,
-            fallback_warning,
-            live_branch_tracked,
-            live_branch_db_path,
-            live_branch_db_exists,
-            nearest_tracked_ancestor,
-            nearest_tracked_ancestor_db_path,
-            nearest_tracked_ancestor_db_exists,
-            tracked_branch_count: branches.len(),
-            branches,
-            warnings,
-        }
-    }
-
     pub fn branch_diagnostics(&self) -> BranchDiagnostics {
-        Self::build_branch_diagnostics(
+        tracedecay_application::tracedecay::build_branch_diagnostics(
             &self.project_root,
             &self.store_layout.data_root,
             self.active_branch.clone(),
